+++ src/index.ts
@@ -1,7 +1,3 @@
-import ApolloProvider from './ApolloProvider';
-import graphql, { withApollo } from './graphql';
+export * from './browser';
 
-// expose easy way to join queries from redux
-import { compose } from 'redux';
-
-export { ApolloProvider, graphql, withApollo, compose };
+export { getDataFromTree, renderToStringWithData } from './server';